+++ packages/Webkul/BookingProduct/src/Providers/ModuleServiceProvider.php
@@ -2,9 +2,9 @@
 
 namespace Webkul\BookingProduct\Providers;
 
-use Konekt\Concord\BaseModuleServiceProvider;
+use Webkul\Core\Providers\CoreModuleServiceProvider;
 
-class ModuleServiceProvider extends BaseModuleServiceProvider
+class ModuleServiceProvider extends CoreModuleServiceProvider
 {
     protected $models = [
         \Webkul\BookingProduct\Models\BookingProduct::class,
@@ -16,25 +16,4 @@ class ModuleServiceProvider extends BaseModuleServiceProvider
         \Webkul\BookingProduct\Models\BookingProductTableSlot::class,
         \Webkul\BookingProduct\Models\Booking::class,
     ];
-
-    public function boot()
-    {
-        if ($this->areMigrationsEnabled()) {
-            $this->registerMigrations();
-        }
-
-        if ($this->areModelsEnabled()) {
-            $this->registerModels();
-            $this->registerEnums();
-            $this->registerRequestTypes();
-        }
-
-        if ($this->areViewsEnabled()) {
-            $this->registerViews();
-        }
-
-        if ($routes = $this->config('routes', true)) {
-            $this->registerRoutes($routes);
-        }
-    }
 }
