+++ packages/Webkul/CatalogRule/src/Providers/ModuleServiceProvider.php
@@ -2,34 +2,13 @@
 
 namespace Webkul\CatalogRule\Providers;
 
-use Konekt\Concord\BaseModuleServiceProvider;
+use Webkul\Core\Providers\CoreModuleServiceProvider;
 
-class ModuleServiceProvider extends BaseModuleServiceProvider
+class ModuleServiceProvider extends CoreModuleServiceProvider
 {
     protected $models = [
         \Webkul\CatalogRule\Models\CatalogRule::class,
         \Webkul\CatalogRule\Models\CatalogRuleProduct::class,
         \Webkul\CatalogRule\Models\CatalogRuleProductPrice::class
     ];
-
-    public function boot()
-    {
-        if ($this->areMigrationsEnabled()) {
-            $this->registerMigrations();
-        }
-
-        if ($this->areModelsEnabled()) {
-            $this->registerModels();
-            $this->registerEnums();
-            $this->registerRequestTypes();
-        }
-
-        if ($this->areViewsEnabled()) {
-            $this->registerViews();
-        }
-
-        if ($routes = $this->config('routes', true)) {
-            $this->registerRoutes($routes);
-        }
-    }
 }
