+++ packages/Webkul/Checkout/src/Providers/ModuleServiceProvider.php
@@ -2,9 +2,9 @@
 
 namespace Webkul\Checkout\Providers;
 
-use Konekt\Concord\BaseModuleServiceProvider;
+use Webkul\Core\Providers\CoreModuleServiceProvider;
 
-class ModuleServiceProvider extends BaseModuleServiceProvider
+class ModuleServiceProvider extends CoreModuleServiceProvider
 {
     protected $models = [
         \Webkul\Checkout\Models\Cart::class,
@@ -13,25 +13,4 @@ class ModuleServiceProvider extends BaseModuleServiceProvider
         \Webkul\Checkout\Models\CartPayment::class,
         \Webkul\Checkout\Models\CartShippingRate::class,
     ];
-
-    public function boot()
-    {
-        if ($this->areMigrationsEnabled()) {
-            $this->registerMigrations();
-        }
-
-        if ($this->areModelsEnabled()) {
-            $this->registerModels();
-            $this->registerEnums();
-            $this->registerRequestTypes();
-        }
-
-        if ($this->areViewsEnabled()) {
-            $this->registerViews();
-        }
-
-        if ($routes = $this->config('routes', true)) {
-            $this->registerRoutes($routes);
-        }
-    }
 }
