+++ packages/Webkul/Core/src/Providers/CoreModuleServiceProvider.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace Webkul\Core\Providers;
+
+use Konekt\Concord\BaseModuleServiceProvider;
+
+class CoreModuleServiceProvider extends BaseModuleServiceProvider
+{
+    public function boot()
+    {
+        if ($this->areMigrationsEnabled()) {
+            $this->registerMigrations();
+        }
+
+        if ($this->areModelsEnabled()) {
+            $this->registerModels();
+            $this->registerEnums();
+            $this->registerRequestTypes();
+        }
+
+        if ($this->areViewsEnabled()) {
+            $this->registerViews();
+        }
+
+        if ($routes = $this->config('routes', true)) {
+            $this->registerRoutes($routes);
+        }
+    }
+}
