+++ packages/Webkul/Core/src/Providers/ModuleServiceProvider.php
@@ -2,9 +2,7 @@
 
 namespace Webkul\Core\Providers;
 
-use Konekt\Concord\BaseModuleServiceProvider;
-
-class ModuleServiceProvider extends BaseModuleServiceProvider
+class ModuleServiceProvider extends CoreModuleServiceProvider
 {
     protected $models = [
         \Webkul\Core\Models\Channel::class,
@@ -19,25 +17,4 @@ class ModuleServiceProvider extends BaseModuleServiceProvider
         \Webkul\Core\Models\Slider::class,
         \Webkul\Core\Models\SubscribersList::class,
     ];
-
-    public function boot()
-    {
-        if ($this->areMigrationsEnabled()) {
-            $this->registerMigrations();
-        }
-
-        if ($this->areModelsEnabled()) {
-            $this->registerModels();
-            $this->registerEnums();
-            $this->registerRequestTypes();
-        }
-
-        if ($this->areViewsEnabled()) {
-            $this->registerViews();
-        }
-
-        if ($routes = $this->config('routes', true)) {
-            $this->registerRoutes($routes);
-        }
-    }
 }
