+++ packages/Webkul/Inventory/src/Providers/ModuleServiceProvider.php
@@ -2,32 +2,11 @@
 
 namespace Webkul\Inventory\Providers;
 
-use Konekt\Concord\BaseModuleServiceProvider;
+use Webkul\Core\Providers\CoreModuleServiceProvider;
 
-class ModuleServiceProvider extends BaseModuleServiceProvider
+class ModuleServiceProvider extends CoreModuleServiceProvider
 {
     protected $models = [
         \Webkul\Inventory\Models\InventorySource::class,
     ];
-
-    public function boot()
-    {
-        if ($this->areMigrationsEnabled()) {
-            $this->registerMigrations();
-        }
-
-        if ($this->areModelsEnabled()) {
-            $this->registerModels();
-            $this->registerEnums();
-            $this->registerRequestTypes();
-        }
-
-        if ($this->areViewsEnabled()) {
-            $this->registerViews();
-        }
-
-        if ($routes = $this->config('routes', true)) {
-            $this->registerRoutes($routes);
-        }
-    }
 }
