+++ packages/Webkul/Sales/src/Providers/ModuleServiceProvider.php
@@ -2,9 +2,9 @@
 
 namespace Webkul\Sales\Providers;
 
-use Konekt\Concord\BaseModuleServiceProvider;
+use Webkul\Core\Providers\CoreModuleServiceProvider;
 
-class ModuleServiceProvider extends BaseModuleServiceProvider
+class ModuleServiceProvider extends CoreModuleServiceProvider
 {
     protected $models = [
         \Webkul\Sales\Models\Order::class,
@@ -20,25 +20,4 @@ class ModuleServiceProvider extends BaseModuleServiceProvider
         \Webkul\Sales\Models\Refund::class,
         \Webkul\Sales\Models\RefundItem::class,
     ];
-
-    public function boot()
-    {
-        if ($this->areMigrationsEnabled()) {
-            $this->registerMigrations();
-        }
-
-        if ($this->areModelsEnabled()) {
-            $this->registerModels();
-            $this->registerEnums();
-            $this->registerRequestTypes();
-        }
-
-        if ($this->areViewsEnabled()) {
-            $this->registerViews();
-        }
-
-        if ($routes = $this->config('routes', true)) {
-            $this->registerRoutes($routes);
-        }
-    }
 }
